app/requirements.txt
requests
kafka-python
psycopg2-binary
sentry-sdk
